requirements.txt
# Bishop Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
openai-whisper>=0.5.0
PyYAML>=6.0
numpy>=1.22.0
requests>=2.28.0
pillow>=9.0.0
emoji>=2.0.0
pydub>=0.25.1
soundfile>=0.11.0
ffmpeg-python>=0.2.0
SpeechRecognition>=3.8.1
pytz>=2022.1
aiohttp>=3.8.1